+++ Layer3/isis/isis_pkt.h
@@ -91,7 +91,7 @@ void
 isis_print_lsp_pkt_cbk(event_dispatcher_t *ev_dis, void *arg, size_t arg_size);
 
 void
-isis_print_hello_pkt_cbk(event_dispatcher_t *ev_dis, void *arg, size_t arg_size);
+isis_print_hello_pkt_cbk(event_dispatcher_t *ev_dis, void *arg, unsigned int arg_size);
 
 void
 isis_cancel_lsp_pkt_generation_task(node_t *node);
